Bazel initial commit PR: libff
* This commit contains libff changes only. It does not include changes
to the embedded ate-pairing and xbyak repos.
* It follows that libff targets will not build until the related
ate-pairing an xbyak PRs have been merged.
