+++ bzl/external/openmp/BUILD.bazel
@@ -1,0 +1,5 @@
+alias(
+    name = "openmp",
+    actual = "@libff//bzl/external/openmp",
+    visibility = ["//visibility:public"],
+)
