+++ bzl/external/openssl/BUILD.bazel
@@ -1,0 +1,5 @@
+alias(
+    name = "openssl",
+    actual = "@libff//bzl/external/openssl",
+    visibility = ["//visibility:public"],
+)
