+++ bzl/obazl.edn
@@ -1,0 +1,33 @@
+{
+ :workspace "snarky"
+ :ppx {:shared-pkg "//bzl/ppx"}
+ :exclusions #{
+               "bench/dune"                                   ;; hand-tuned ppx arg: -inline-test-lib
+               "h_list/dune"                                  ;; singleton
+               "h_list/ppx/dune"                              ;; singleton
+               "snarky_universe/dune"                         ;; hand-tuned
+               "src/camlsnark_c/dune"                         ;; genrules
+               "src/camlsnark_c/cpp_string/dune"              ;; singleton
+               "src/camlsnark_c/cpp_vector/dune"              ;; hand-tuned: Vector_ffi_bindings dep
+
+               "src/dune"                                     ;; cc interop
+               "src/camlsnark_c/cpp_vector/gen/dune"          ;; cc interop
+               "src/camlsnark_c/libsnark_bindings/dune"       ;; cc interop
+
+               ;; not yet supported: meja
+               "meja/dune"
+               "meja/ocaml/dune"
+               "meja/ocaml/of_ocaml/dune"
+               "meja/src/dune"
+               "meja/src/compiler_internals/dune"
+
+               ;; unused
+               "dune"                                         ;; not used
+               "src/camlsnark_c/groth16_gpu_prover/dune"      ;; not used
+               "src/camlsnark_c/groth16_gpu_prover/test/dune" ;; not used
+               "src/camlsnark_c/snark_caml_bn128/dune"        ;; not used
+               "src/camlsnark_c/snark_caml_common/dune"       ;; not used
+               "src/camlsnark_c/snark_caml_mnt298/dune"       ;; not used
+               "src/camlsnark_c/snark_caml_mnt753/dune"       ;; not used
+               }
+ }
