+++ src/camlsnark_c/libsnark-caml/depends/libff/.bazelrc
@@ -1,0 +1,6 @@
+# enable platform-based toolchain resolution:
+build --incompatible_enable_cc_toolchain_resolution
+
+build --color=yes
+
+try-import user.bazelrc
