+++ src/camlsnark_c/libsnark-caml/depends/libff/.gitignore
@@ -1,14 +1,8 @@
 .bazel
 bazel-*
 user.bazelrc
-*.log
-# obazl tool:
-.cpcache
-deps.edn
 
 .merlin
 _build
 *.install
 *:*
-
-*~
