+++ src/camlsnark_c/libsnark-caml/depends/libff/bzl/external/openmp/BUILD.bazel
@@ -1,0 +1,13 @@
+load("@rules_foreign_cc//tools/build_defs:cmake.bzl", "cmake_external")
+
+## Use libgomp for linux?
+
+cmake_external(
+    name = "openmp",
+    lib_source = "@openmp//:all",
+    shared_libraries = select({
+        "//bzl/host:macos": ["libomp.dylib"],
+        "//bzl/host:linux": ["libomp.so"],
+    }, no_match_error = "OpenMP: unsupported platform.  Linux and MacOS only."),
+    visibility = ["//visibility:public"],
+)
