+++ src/camlsnark_c/libsnark-caml/depends/libff/bzl/external/openssl/BUILD.bazel
@@ -1,0 +1,35 @@
+load("@rules_foreign_cc//tools/build_defs:configure.bzl", "configure_make")
+
+## MacOS: make sure the native clang tools (e.g. 'ar') are not shadowed by e.g. homebrew binutils.
+
+configure_make(
+    name = "openssl",
+    configure_command = "config",
+    configure_in_place = True,
+    configure_env_vars = select({
+        "//bzl/host:macos": {
+            "AR": "",
+            "CXXFLAGS" : "-lc++ -m64 -fomit-frame-pointer -fPIC -DPIC",
+        },
+        "//bzl/host:linux_x64": {
+            "CFLAGS" : "-fPIC -DPIC -m64",
+            "CXXFLAGS" : "-lstdc++ -m64 -fomit-frame-pointer -fPIC -DPIC",
+        }
+    }, no_match_error = "Unsupported platform.  Linux and MacOS only."),
+    configure_options = select({
+        "//bzl/host:macos": ["ARFLAGS=r"],
+        "//conditions:default": []
+    }),
+    lib_source = "@openssl//:all",
+    out_lib_dir = "lib",
+    shared_libraries = select({
+        "//bzl/host:macos": ["libssl.dylib", "libcrypto.dylib"],
+        "//bzl/host:linux": ["libssl.so", "libcrypto.so"],
+    }, no_match_error = "Openssl: Unsupported platform.  Linux and MacOS only."),
+    static_libraries = [
+        "libssl.a",
+        "libcrypto.a",
+    ],
+    visibility = ["//visibility:public"],
+)
+
