+++ src/camlsnark_c/libsnark-caml/depends/libff/bzl/tools/aliaslog
@@ -1,0 +1,3 @@
+# source me
+
+alias "bl=less `bazel info command_log`"
