+++ src/camlsnark_c/libsnark-caml/depends/libff/libff/CMakeLists.txt
@@ -56,8 +56,8 @@ add_library(
   algebra/curves/mnt753/mnt6753/mnt6753_init.cpp
   algebra/curves/mnt753/mnt6753/mnt6753_pairing.cpp
   algebra/curves/mnt753/mnt6753/mnt6753_pp.cpp
-  common/common.cpp
   common/double.cpp
+  common/logging.cpp
   common/profiling.cpp
   common/utils.cpp
