+++ src/camlsnark_c/libsnark-caml/depends/libff/libff/algebra/curves/edwards/BUILD.bazel
@@ -1,0 +1,55 @@
+load("@rules_cc//cc:defs.bzl", "cc_library")
+load("@libff//bzl/config:vars.bzl", "CPPFLAGS", "CXXFLAGS",
+     "ALWAYSLINK", "LINKSTATIC",
+     "DDEBUG", "DNO_PT_COMPRESSION", "DPROFILE_OP_COUNTS")
+
+cc_library(
+    name = "edwards",
+    srcs = [
+        "edwards_g1.cpp",
+        "edwards_g2.cpp",
+        "edwards_init.cpp",
+        "edwards_pairing.cpp",
+    ],
+    hdrs = [
+        "edwards_g1.hpp",
+        "edwards_g2.hpp",
+        "edwards_init.hpp",
+        "edwards_pairing.hpp",
+    ],
+    copts = ["-I."] + CPPFLAGS + CXXFLAGS,
+    local_defines = DDEBUG + DNO_PT_COMPRESSION + DPROFILE_OP_COUNTS,
+    include_prefix = "libff/algebra/curves/edwards",
+    linkstatic = LINKSTATIC,
+    alwayslink = ALWAYSLINK,
+    visibility = ["//visibility:public"],
+    deps = [
+        "//libff/common:profiling",
+        "//libff/algebra/curves",
+        "//libff/algebra/curves/edwards:public_params",
+        "//libff/algebra/fields",
+    ],
+)
+
+cc_library(
+    name = "public_params",
+    srcs = [
+        "edwards_pp.cpp",
+    ],
+    hdrs = [
+        "edwards_g1.hpp",
+        "edwards_g2.hpp",
+        "edwards_init.hpp",
+        "edwards_pairing.hpp",
+        "edwards_pp.hpp",
+    ],
+    copts = ["-I."] + CPPFLAGS + CXXFLAGS,
+    local_defines = DDEBUG + DNO_PT_COMPRESSION + DPROFILE_OP_COUNTS,
+    include_prefix = "libff/algebra/curves/edwards",
+    linkstatic = LINKSTATIC,
+    visibility = ["//visibility:public"],
+    deps = [
+        "//libff/algebra/curves",
+        "//libff/algebra/fields",
+    ],
+)
