+++ src/camlsnark_c/libsnark-caml/depends/libff/libff/algebra/curves/mnt/BUILD.bazel
@@ -1,0 +1,15 @@
+load("@rules_cc//cc:defs.bzl", "cc_library")
+load("@libff//bzl/config:vars.bzl", "CPPFLAGS", "CXXFLAGS",
+     "DDEBUG", "DNO_PT_COMPRESSION")
+
+cc_library(
+    name = "mnt",
+    srcs = ["mnt46_common.cpp"],
+    hdrs = ["mnt46_common.hpp"],
+    copts = ["-I."] + CPPFLAGS + CXXFLAGS,
+    include_prefix = "libff/algebra/curves/mnt",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//libff/algebra/fields",
+    ],
+)
