+++ src/camlsnark_c/libsnark-caml/depends/libff/libff/algebra/curves/mnt/mnt4/BUILD.bazel
@@ -1,0 +1,55 @@
+load("@rules_cc//cc:defs.bzl", "cc_library")
+load("@libff//bzl/config:vars.bzl", "CPPFLAGS", "CXXFLAGS",
+     "ALWAYSLINK", "LINKSTATIC",
+     "DDEBUG", "DNO_PT_COMPRESSION", "DPROFILE_OP_COUNTS")
+
+cc_library(
+    name = "mnt4",
+    srcs = [
+        "mnt4_g1.cpp",
+        "mnt4_g2.cpp",
+        "mnt4_init.cpp",
+        "mnt4_pairing.cpp",
+        "mnt4_pp.cpp",
+    ],
+    hdrs = [
+        "mnt4_g1.hpp",
+        "mnt4_g2.hpp",
+        "mnt4_init.hpp",
+        "mnt4_pairing.hpp",
+        "mnt4_pp.hpp",
+    ],
+    copts = ["-I."] + CPPFLAGS + CXXFLAGS,
+    local_defines = DDEBUG + DNO_PT_COMPRESSION + DPROFILE_OP_COUNTS,
+    include_prefix = "libff/algebra/curves/mnt/mnt4",
+    linkstatic = LINKSTATIC,
+    # alwayslink = ALWAYSLINK,
+    visibility = ["//visibility:public"],
+    deps = [
+        "//libff/common:profiling",
+        "//libff/algebra/curves",
+        "//libff/algebra/curves/mnt",
+        ":public_params",
+    ],
+)
+
+cc_library(
+    name = "public_params",
+    srcs = ["mnt4_pp.cpp"],
+    hdrs = [
+        "mnt4_g1.hpp",
+        "mnt4_g2.hpp",
+        "mnt4_init.hpp",
+        "mnt4_pairing.hpp",
+        "mnt4_pp.hpp",
+    ],
+    copts = ["-I."] + CPPFLAGS + CXXFLAGS,
+    local_defines = DDEBUG + DNO_PT_COMPRESSION + DPROFILE_OP_COUNTS,
+    include_prefix = "libff/algebra/curves/mnt/mnt4",
+    linkstatic = LINKSTATIC,
+    visibility = ["//visibility:public"],
+    deps = [
+        "//libff/algebra/curves",
+        "//libff/algebra/curves/mnt",
+    ],
+)
