+++ src/camlsnark_c/libsnark-caml/depends/libff/libff/algebra/curves/mnt753/mnt6753/BUILD.bazel
@@ -1,0 +1,57 @@
+load("@rules_cc//cc:defs.bzl", "cc_library")
+load("@libff//bzl/config:vars.bzl", "CPPFLAGS", "CXXFLAGS",
+     "ALWAYSLINK", "LINKSTATIC",
+     "DDEBUG", "DNO_PT_COMPRESSION", "DPROFILE_OP_COUNTS")
+
+cc_library(
+    name = "mnt6753",
+    srcs = [
+        "mnt6753_g1.cpp",
+        "mnt6753_g2.cpp",
+        "mnt6753_init.cpp",
+        "mnt6753_pairing.cpp",
+        "mnt6753_pp.cpp",
+    ],
+    hdrs = glob([
+        "mnt6753_g1.hpp",
+        "mnt6753_g2.hpp",
+        "mnt6753_init.hpp",
+        "mnt6753_pairing.hpp",
+        "mnt6753_pp.hpp",
+    ]),
+    copts = ["-I."] + CPPFLAGS + CXXFLAGS,
+    local_defines = DDEBUG + DNO_PT_COMPRESSION + DPROFILE_OP_COUNTS,
+    include_prefix = "libff/algebra/curves/mnt753/mnt6753",
+    linkstatic = LINKSTATIC,
+    # alwayslink = ALWAYSLINK,
+    visibility = ["//visibility:public"],
+    deps = [
+        "//libff/common:profiling",
+        "//libff/algebra/curves",
+        "//libff/algebra/curves/mnt",
+        "//libff/algebra/curves/mnt753",
+        ":public_params",
+    ],
+)
+
+cc_library(
+    name = "public_params",
+    srcs = ["mnt6753_pp.cpp"],
+    hdrs = glob([
+        "mnt6753_g1.hpp",
+        "mnt6753_g2.hpp",
+        "mnt6753_init.hpp",
+        "mnt6753_pairing.hpp",
+        "mnt6753_pp.hpp",
+    ]),
+    copts = ["-I."] + CPPFLAGS + CXXFLAGS,
+    local_defines = DDEBUG + DNO_PT_COMPRESSION + DPROFILE_OP_COUNTS,
+    include_prefix = "libff/algebra/curves/mnt753/mnt6753",
+    linkstatic = LINKSTATIC,
+    visibility = ["//visibility:public"],
+    deps = [
+        "//libff/algebra/curves",
+        "//libff/algebra/fields",
+        "//libff/algebra/curves/mnt753",
+    ],
+)
