+++ src/camlsnark_c/libsnark-caml/depends/libff/libff/common/default_types/BUILD.bazel
@@ -1,0 +1,21 @@
+load("@rules_cc//cc:defs.bzl", "cc_library")
+load("@libff//bzl/config:vars.bzl", "CPPFLAGS",
+     "DBINARY_OUTPUT", "DCURVE", "DDEBUG", "DMULTICORE", "DNO_PROCPS")
+
+cc_library(
+    name = "default_types",
+    hdrs = [
+        "ec_pp.hpp",
+    ],
+    copts = ["-I.", "-Iexternal"] + CPPFLAGS,
+    defines = DCURVE,
+    visibility = ["//visibility:public"],
+    include_prefix = "libff/common/default_types",
+    deps = select({
+        "@//bzl/config:enable_curve_bn128": ["//libff/algebra/curves/bn128"],
+        "@//bzl/config:enable_curve_alt_bn128": ["//libff/algebra/curves/alt_bn128:public_params"],
+        "@//bzl/config:enable_curve_edwards": ["//libff/algebra/curves/edwards:public_params"],
+        "@//bzl/config:enable_curve_mnt4": ["//libff/algebra/curves/mnt/mnt4:public_params"],
+        "@//bzl/config:enable_curve_mnt6": ["//libff/algebra/curves/mnt/mnt6:public_params"],
+    }, no_match_error = "ERROR: no curve selected.  Try --//:curve=<curve>")
+)
